backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.35
alembic==1.13.3
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Validation & Settings
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# OAuth
authlib==1.3.2
httpx==0.27.2

# Redis & Caching
redis==5.1.1
hiredis==3.0.0

# Task Queue
celery==5.4.0

# Email
resend==2.4.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.2

# Code Quality
ruff==0.6.9
mypy==1.11.2

# Type Stubs
types-redis==4.6.0.20240903
types-passlib==1.7.7.20240819

# Utilities
python-dotenv==1.0.1

bcrypt==4.0.1